safari-600.7-branch/LayoutTests/svg/custom/invalid-transforms.svg

<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1 Basic//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11-basic.dtd">
<svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" viewBox="0 0 100 100" onload="runTest()">
    <script type="text/ecmascript"><![CDATA[
    function runTest() {
        var element = document.getElementById('test');
        element.setAttribute("transform", "rotate(45) scale(1,2,3)");
    }
    ]]></script>
    <!-- The current behavior is to ignore invalid transforms (and any transforms after).
         We might choose to enter an "error state" on an invlalid transform in the future. -->
    <rect width="100" height="100" fill="red"/>
    <rect id="test" width="50" height="50" transform="scale(2,2) scale(1,2,3) translate(50,50)" fill="green"/>
</svg>
