third_party/WebKit/LayoutTests/svg/custom/invalid-transforms-expected.txt

CONSOLE ERROR: line 13: Error: <rect> attribute transform: Expected ')', "…e(2,2) scale(1,2,3) translate(50…".
CONSOLE ERROR: line 7: Error: <rect> attribute transform: Expected ')', "…te(45) scale(1,2,3)".
